requirements.txt
python-garminconnect>=0.2.22
matplotlib>=3.8.0
anthropic>=0.40.0
fastembed>=0.3.0
chromadb>=0.5.0
